+++ src/main/kotlin/graphql/dataTransferObjects/StarredMessage.kt
@@ -2,10 +2,7 @@
 
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
-import com.neelkamath.omniChatBackend.db.ForwardPagination
 import com.neelkamath.omniChatBackend.db.tables.Messages
-import com.neelkamath.omniChatBackend.graphql.routing.MessageState
-import graphql.schema.DataFetchingEnvironment
 import java.time.LocalDateTime
 
 sealed interface StarredMessage {
@@ -16,14 +13,7 @@ sealed interface StarredMessage {
 
     fun getSender(): Account = Account(Messages.readSenderId(id))
 
-    fun getState(): MessageState = Messages.readState(id)
-
     fun getSent(): LocalDateTime = Messages.readSent(id)
-
-    fun getStatuses(env: DataFetchingEnvironment): MessageDateTimeStatusConnection {
-        val pagination = ForwardPagination(env.getArgument("first"), env.getArgument("after"))
-        return MessageDateTimeStatusConnection(id, pagination)
-    }
 
     fun getContext(): MessageContext = MessageContext(id)