+++ src/test/kotlin/graphql/dataTransferObjects/MessageTest.kt
@@ -6,10 +6,7 @@ import com.neelkamath.omniChatBackend.createVerifiedUsers
 import com.neelkamath.omniChatBackend.db.ForwardPagination
 import com.neelkamath.omniChatBackend.db.tables.*
 import com.neelkamath.omniChatBackend.graphql.engine.executeGraphQlViaEngine
-import com.neelkamath.omniChatBackend.graphql.routing.Cursor
 import com.neelkamath.omniChatBackend.graphql.routing.GroupChatPublicity
-import com.neelkamath.omniChatBackend.graphql.routing.MessageStatus
-import com.neelkamath.omniChatBackend.slice
 import com.neelkamath.omniChatBackend.testingObjectMapper
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.extension.ExtendWith
@@ -18,23 +15,14 @@ import kotlin.test.assertEquals
 
 @ExtendWith(DbExtension::class)
 class MessageTest {
-    private data class ReadMessageResponse(val hasStar: Boolean, val statuses: Statuses) {
-        data class Statuses(val edges: List<Edge>) {
-            data class Edge(val cursor: Cursor)
-        }
-    }
+    private data class ReadMessageResponse(val hasStar: Boolean)
 
     private fun readMessage(userId: Int?, messageId: Int, pagination: ForwardPagination? = null): ReadMessageResponse {
         val data = executeGraphQlViaEngine(
             """
-            query ReadMessage(${"$"}messageId: Int!, ${"$"}first: Int, ${"$"}after: Cursor) {
+            query ReadMessage(${"$"}messageId: Int!) {
                 readMessage(messageId: ${"$"}messageId) {
                     ... on Message {
-                        statuses(first: ${"$"}first, after: ${"$"}after) {
-                            edges {
-                                cursor
-                            }
-                        }
                         hasStar
                     }
                 }
@@ -44,23 +32,6 @@ class MessageTest {
             userId,
         ).data!!["readMessage"] as Map<*, *>
         return testingObjectMapper.convertValue(data)
-    }
-
-    @Nested
-    inner class GetStatuses {
-        @Test
-        fun `Statuses must be paginated`() {
-            val adminId = createVerifiedUsers(1).first().userId
-            val userIdList = createVerifiedUsers(10).map { it.userId }
-            val chatId = GroupChats.create(setOf(adminId), userIdList)
-            val messageId = Messages.message(adminId, chatId)
-            userIdList.forEach { MessageStatuses.create(it, messageId, MessageStatus.DELIVERED) }
-            val statusIdList = MessageStatuses.readIdList(messageId)
-            val index = 4
-            val pagination = ForwardPagination(first = 3, after = statusIdList.elementAt(index))
-            val actual = readMessage(adminId, messageId, pagination).statuses.edges.map { it.cursor }
-            assertEquals(statusIdList.slice(index + 1..index + pagination.first!!).toList(), actual)
-        }
     }
 
     @Nested